Move resource lifecycle helpers to a common utils module.
