+++ src/client/common/asyncDisposableRegistry.ts
@@ -2,26 +2,9 @@
 // Licensed under the MIT License.
 'use strict';
 import { injectable } from 'inversify';
-import { IAsyncDisposable, IAsyncDisposableRegistry, IDisposable } from './types';
+import { IAsyncDisposableRegistry } from './types';
+import { Disposables } from './utils/resourceLifecycle';
 
 // List of disposables that need to run a promise.
 @injectable()
-export class AsyncDisposableRegistry implements IAsyncDisposableRegistry {
-    private _list: (IDisposable | IAsyncDisposable)[] = [];
-
-    public async dispose(): Promise<void> {
-        const promises = this._list.map((l) => l.dispose());
-        await Promise.all(promises);
-        this._list = [];
-    }
-
-    public push(disposable?: IDisposable | IAsyncDisposable) {
-        if (disposable) {
-            this._list.push(disposable);
-        }
-    }
-
-    public get list(): (IDisposable | IAsyncDisposable)[] {
-        return this._list;
-    }
-}
+export class AsyncDisposableRegistry extends Disposables implements IAsyncDisposableRegistry {}
