+++ src/client/common/platform/fileSystemWatcher.ts
@@ -7,8 +7,7 @@ import * as chokidar from 'chokidar';
 import * as path from 'path';
 import { RelativePattern, workspace } from 'vscode';
 import { traceError, traceVerbose, traceWarning } from '../logger';
-import { DisposableRegistry } from '../syncDisposableRegistry';
-import { IDisposable } from '../types';
+import { IDisposable } from '../utils/resourceLifecycle';
 import { normCasePath } from './fs-paths';
 
 /**
@@ -25,7 +24,7 @@ export function watchLocationForPattern(
     baseDir: string,
     pattern: string,
     callback: (type: FileChangeType, absPath: string) => void
-): IDisposable {
+): IDisposable[] {
     // Use VSCode API iff base directory to exists within the current workspace folders
     const found = workspace.workspaceFolders?.find((e) => normCasePath(baseDir).startsWith(normCasePath(e.uri.fsPath)));
     if (found) {
@@ -39,9 +38,9 @@ function watchLocationUsingVSCodeAPI(
     baseDir: string,
     pattern: string,
     callback: (type: FileChangeType, absPath: string) => void
-): IDisposable {
+): IDisposable[] {
     const globPattern = new RelativePattern(baseDir, pattern);
-    const disposables = new DisposableRegistry();
+    const disposables: IDisposable[] = [];
     traceVerbose(`Start watching: ${baseDir} with pattern ${pattern} using VSCode API`);
     const watcher = workspace.createFileSystemWatcher(globPattern);
     disposables.push(watcher.onDidCreate((e) => callback(FileChangeType.Created, e.fsPath)));
@@ -54,7 +53,7 @@ function watchLocationUsingChokidar(
     baseDir: string,
     pattern: string,
     callback: (type: FileChangeType, absPath: string) => void
-): IDisposable {
+): IDisposable[] {
     const watcherOpts: chokidar.WatchOptions = {
         cwd: baseDir,
         ignoreInitial: true,
@@ -129,5 +128,7 @@ function watchLocationUsingChokidar(
         }
     });
 
-    return { dispose: () => stopWatcher().ignoreErrors() };
+    return [
+        { dispose: () => stopWatcher().ignoreErrors() },
+    ];
 }
