+++ src/client/pythonEnvironments/base/locators/common/resourceBasedLocator.ts
@@ -1,66 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { DisposableRegistry } from '../../../../common/syncDisposableRegistry';
-import { logWarning } from '../../../../logging';
-import { Locator, } from '../../locator';
-
-// TODO: Move `Activatable` and friends to an appropriate location.  (`src/client/common/utils/resources`?)
-
-export interface IDisposable {
-    dispose(): void | Promise<void>;
-}
-
-export interface IActivatable {
-    activate(): Promise<void>;
-    dispose(): Promise<void>;
-    readonly active: boolean;
-}
-
-class Activatable implements IActivatable {
-    private pending = false;
-
-    private activated = false;
-
-    private readonly disposables: IDisposable[] = [];
-
-    constructor(
-        private readonly do_activation: () => Promise<IDisposable[]>,
-    ) {}
-
-    public async activate(): Promise<void> {
-        if (this.pending || this.activated) {
-            return;
-        }
-        this.pending = true;
-        const disposables = await this.do_activation();
-        this.disposables.push(...disposables);
-        this.pending = false;
-        this.activated = true;
-    }
-
-    public async dispose(): Promise<void> {
-        if (this.pending || !this.activated) {
-            return;
-        }
-        this.pending = true;
-        this.activated = false;
-        await Promise.all(
-            this.disposables.map(async (d, index) => {
-                try {
-                    await d.dispose();
-                } catch (err) {
-                    logWarning(`dispose() #${index} failed (${err})`);
-                }
-            }),
-        );
-        this.pending = false;
-    }
-
-    public get active(): boolean {
-        return this.activated;
-    }
-}
+import { Activatable, Disposables, IActivatable, IDisposable } from '../../../../common/utils/resourceLifecycle';
+import { Locator } from '../../locator';
 
 /**
  * A locator that has resources to be activated and disposed.
@@ -94,7 +36,7 @@ export abstract class ResourceBasedLocator extends Locator implements IActivatab
  * A locator that has things to dispose.
  */
 export abstract class DisposableLocator extends Locator {
-    protected readonly disposables = new DisposableRegistry();
+    protected readonly disposables = new Disposables();
 
     public dispose(): void {
         this.disposables.dispose();
